04_Deployement/dashboard/requirements.txt
numpy<2
pandas==2.1.4
streamlit==1.35.0
plotly==5.22.0
psycopg2-binary==2.9.9
confluent-kafka==2.3.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
